requirements.txt
archivebox>=0.7.2
click>=8.2.2
PyYAML>=6.0.2